Fix CSV URL and request payload handling
Previously, we thought that the payload itself was the value of the jobParam query parameter, but it appears that the payload includes this as well.  We need to decode the Rison value before modifying the underlying bool filter and then encode it again when injecting them into the request payload on the server itself.  This changeset attempts to fix the order of things to make this work.